Replace classes with aliases.
Modeling all of ActivityPub's types as model classes is expensive,
because each model class comes with machinery for type-specific CRUD
operations, even when their behavior is identical to the base type.
Crystal aliases are not introspectable at compile-time so enumerate
aliases in a constant for each base model class. This change reduces
executable size by ~16.9%.